Init project
The purpose of this project is to do web scraping of the official FFXIV forums then analyse the forum posts before presenting the results in a dashboard/webpage format. Added classes for saving the data in a SQLite database. Added webscraping of the healer forum (front page). Started on text analysis and machine learning.